+++ .gitignore
@@ -1,0 +1,4 @@
+/nbproject/private/
+/build/
+/dist/
+/nbproject/
