+++ .gradle/2.2.1/taskArtifacts/cache.properties
@@ -1,0 +1,1 @@
+#Mon Mar 30 21:30:49 CEST 2015
