+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Feb 27 12:47:11 CET 2015
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.2.1-bin.zip
