+++ src/main/java/com/chrisnewland/demofx/PreCalc.java
